FEAT: Add status slash command
The /status command allows the bot to check the current IP the bot uses and notify the discord if it has changed from the expected IP. This is useful for users who self-host on their own network. If the palworld server is shutdown, it will also tell the discord users that the server is unavailable.